Support docker registry auth
This adds a new parameter to support docker registry authentication.
This allows us to avoid rate limiting issues in our CI environment that
we've been hitting frequently due to unauthenticated registry calls.

The parameter is a JSON object of the same form of the standard docker
config json file, which means you can directly pass the config file or
pass a manually created object.

I've also added the `--progress plain` flag to the build, to make the
ops output friendlier in CI environments.